requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
python-dotenv==1.0.0
pydantic>=2.6.0
jsonschema==4.20.0
colorama==0.4.6
rich==13.7.0
pandas>=2.2.0
